+++ CHANGELOG.md
@@ -22,8 +22,10 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Support Jackson 3.x release line ([#1810](https://github.com/opensearch-project/opensearch-java/pull/1810))
 - Added `equals()` and `hashCode()` implementations to `FieldValue` ([#1998](https://github.com/opensearch-project/opensearch-java/pull/1998))
 - Add document lifecycle guide and runnable sample ([#2017](https://github.com/opensearch-project/opensearch-java/pull/2017))
+- Add configurable I/O reactor rebuild back-off to `ApacheHttpClient5Transport` ([#1969](https://github.com/opensearch-project/opensearch-java/issues/1969))
 
 ### Fixed
+- Recover `ApacheHttpClient5Transport` from an I/O reactor shutdown instead of failing every subsequent request permanently ([#1969](https://github.com/opensearch-project/opensearch-java/issues/1969))
 
 ## [Unreleased 3.x]
 ### Added